50_libraries/5_NATS/project-1/pyproject.toml
[tool.poetry]
name = "test-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "test_1"}]

[tool.poetry.dependencies]
python = "^3.11.0"
nats-py = "^2.6.0"

[tool.poetry.group.dev.dependencies]
rope = "^1.7.0"
black = "^23.1.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
